requirements.txt
openai==1.30.3
Flask==3.0.3
python-dotenv==1.0.1